Rename league statistics files to adhere to convention
