+++ lib/league_statistics.rb
